pyproject.toml
[project]
name = "openagent"
version = "2.0.0"
description = "OpenAgent is a framework for building AI applications leveraging the power of blockchains and RSS3 Network."
requires-python = ">=3.12,<4.0"
dependencies = [
    "anthropic>=0.42.0",
    "fastapi>=0.115.6",
    "google-generativeai>=0.8.3",
    "loguru",
    "ollama>=0.4.5",
    "openai>=1.59.6",
    "phidata>=2.7.7",
    "python-dotenv>=1.0.1",
    "sqlalchemy-orm>=1.2.10",
    "uvicorn>=0.34.0",
    "yfinance>=0.2.51",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "atlas-provider-sqlalchemy (>=0.2.4,<0.3.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "eth-account (>=0.13.4,<0.14.0)",
    "web3 (>=7.7.0,<8.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "tweepy (>=4.15.0,<5.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "jwt (>=1.3.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
ruff = "^0.9.1"

[tool.ruff]
line-length = 150
indent-width = 4
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint.mccabe]
max-complexity = 7

[tool.ruff.lint]
select = [
    "AIR",
    "B",
    "C90",
    "E",
    "F",
    "FLY",
    "FURB",
    "I",
    "N",
    "PERF",
    "RUF",
    "SIM",
    "UP",
    "W"
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
